feat: createOrder component created, added MantineProvider and Container to App.jsx for styles
Built the starting UI for the CreateOrder component which includes the Vendor selection and Order
Title input, Create Order header, and a Next button. Added this to the App.jsx for viewing in
browser. Also modified the eslintrc file to only warn when there are unused varaibles rather than
throw an error to help speed up development time